updates setup.py to compile Python wrapper with libboost python 3 if available
